+++ lib/common.dart
@@ -5,6 +5,7 @@ export 'package:kakao_flutter_sdk/src/common/api_factory.dart';
 export 'package:kakao_flutter_sdk/src/common/constants.dart';
 export 'package:kakao_flutter_sdk/src/common/kakao_api_exception.dart';
 export 'package:kakao_flutter_sdk/src/common/kakao_auth_exception.dart';
+export 'package:kakao_flutter_sdk/src/common/dapi_exception.dart';
 export 'package:kakao_flutter_sdk/src/common/kakao_context.dart';
 export 'package:kakao_flutter_sdk/src/common/kakao_error.dart';
 export 'package:kakao_flutter_sdk/src/common/util.dart';