+++ lib/src/search/model/search_meta.dart
@@ -4,8 +4,13 @@ part 'search_meta.g.dart';
 
 @JsonSerializable(fieldRename: FieldRename.snake)
 class SearchMeta {
+  /// total number of documents
   int totalCount;
+
+  ///
   int pageableCount;
+
+  /// Whether current page is the end page. Next request can be made with an incremented page number if this flag is false.
   bool isEnd;
 
   SearchMeta(this.totalCount, this.pageableCount, this.isEnd);